Update README.md
The ATM Management System is a tiny project, which is used to access their bank accounts in order to make cash withdrawals. Whenever user want to make withdrawals, they can enter their ATM card and verified PIN ,then user select the withdrawal option and enter the withdrawal amounts and it will display the amount to user withdrawn in the form of 50’s. lOO’s and 500’s. The user also able to perform one or more transactions. Security is the foundation of good ATM system. This system will provide for secure connections between users and the bank servers. The whole process will be automated right from PIN validation to transaction completion.